Extract types info separate types/cors.js file and minor fixes
