+++ v3/src/js/types/views.js
@@ -30,39 +30,3 @@ export type PageRangeDiff = {
 };
 
 export type OnPageChange = PageRangeDiff => void;
-
-/* components/CorsStats */
-
-export type StudentType = number;
-export const NEW_STUDENT = 1;
-export const RETURNING_STUDENT = 2;
-export const GENERAL_ACCOUNT = 4;
-
-// Simplified version of BiddingStat
-export type GroupedBiddingStat = {
-  AcadYear: AcadYear,
-  Faculty: Faculty,
-  Semester: string,
-  StudentType: StudentType,
-  Round: string,
-  Quota: number,
-  Bidders: number,
-  LowestSuccessfulBid: number,
-};
-
-export type BiddingSummary = {
-  [Faculty]: {
-    [StudentType]: {
-      minBid: number,
-      round: string,
-    },
-  }
-};
-
-export type SemesterStats = {
-  quota: number,
-  bids: number,
-  faculties: Set<Faculty>,
-  stats: GroupedBiddingStat[],
-  summary: BiddingSummary,
-};
